QA: Redesign flow_collector.php to support IPFIX
I read through the fields specs here:

https://www.iana.org/assignments/ipfix/ipfix.xhtml

And here:

https://www.cisco.com/en/US/technologies/tk648/tk362/technologies_white_paper09186a00800a3db9.html

And found that there are subtile version differences that need additional inspection beyond the obvious.

So, I aligned the $allfields array to match the iana specification and
then use a function two switch things up when the type can be variable depending on the version.

These are still some field specifications that are undefined, but we are not presently using them anyway.